[agent] peak-finding renames
Thanks for deleting the old peak-finding code!

You'll notice that a lot of names (either function names, class names, or filenames) have the number "2" in them. This made sense when there were two versions of the peak-finding code, but is now undesirable. For example, I'd like to rename GpuPeakFindingKernel2 -> GpuPeakFindingKernel.

Make a list of filenames, class names, and filenames that should be changed, now that there is only one version of the peak-finding code. Each entry in the list should have the form (old name) -> (new name). For example, two entries on your list might be:

 C++ class GpuPeakFindingKernel2 -> C++ class GpuPeakFindingKernel.
 File python_frb/PeakFinder2.py -> python_frb/PeakFinder.py

(I'll ask you to implement these renames in the next task, but for now just decide what renames should be performed.)

=======================

That's great! Please implement this list of changes.