[agent] Move src_lib/time_* into static member functions.
The files

  - time_gpu_dedispersion_kernels.cu (*)
  - time_gpu_lagged_downsampling_kernels.cu (**)

each contain a "top-level" function which is called from 'python -m pirate_frb time', and a bunch of helper classes/functions.

I'd like all of this code (top-level and helpers) moved to

   src_lib/GpuDedispersionKernel.cu
   src_lib/GpuLaggedDownsamplingKernel.cu

and the "top-level" functions to be callable as new static member functions:

  GpuDedispersionKernel::time()
  GpuLaggedDownsamplingKernel::time()

The helper classes/functions shouldn't be externally visible, in order to avoid polluting the namespace. For example, you might put them in anonymous namespaces.

Please modify __main__.py and the pybind11 linkage so that 'python -m pirate_frb time' still works.