[agent] scratch()
Add a function pirate::scratch() to utils.cu, with pybind11 linkage, and modify __main__.py so that the C++ function scratch() is callable with 'python -m pirate_frb scratch'. Comments in the python code, and the help text for the scratch command, should indicate that scratch() is defined in utils.cu.